# server/lib/v3/api.yaml
openapi: 3.0.3
info:
  version: 3.0.0
  title: Midori Wallet API V3
servers:
  - url: https://wallet.astian.org/api/v3
paths:
...
  /csfee:
    get:
      x-eov-operation-id: getCsFee
      x-eov-operation-handler: handlers
      description: Get transaction cs fees for coin
      parameters:
        - $ref: '#/components/parameters/maybeDeviceId'
        - $ref: '#/components/parameters/cryptoId'
      security:
        - deviceSignature: []
        - apiKey: []
      responses:
        '200':
          description: Fees for coin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsFee'
        default:
          $ref: '#/components/responses/GeneralError'
  /details:
    get:
      x-eov-operation-id: getDetails
      x-eov-operation-handler: handlers
      description: Get user details
      parameters:
        - $ref: '#/components/parameters/deviceId'
      security:
        - deviceSignature: []
      responses:
        '200':
          description: User datails
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: '#/components/schemas/Base64String'
                required:
                  - data
        default:
          $ref: '#/components/responses/GeneralError'
    put:
      x-eov-operation-id: setDetails
      x-eov-operation-handler: handlers
      description: Update user details
      parameters:
        - $ref: '#/components/parameters/deviceId'
      security:
        - deviceSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                data:
                  $ref: '#/components/schemas/Base64String'
              required:
                - data
      responses:
        '200':
          description: Update user details result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: '#/components/schemas/Base64String'
                required:
                  - data
        default:
          $ref: '#/components/responses/GeneralError'
...
  /download/{distribution}/{arch}:
    get:
      x-eov-operation-id: downloadApp
      x-eov-operation-handler: handlers
      description: Get release
      parameters:
        - $ref: '#/components/parameters/distribution'
        - $ref: '#/components/parameters/arch'
      security: []
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '302':
          description: Redirect to file
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RedirectString'
          headers:
            Location:
              schema:
                $ref: '#/components/schemas/UrlString'
        default:
          $ref: '#/components/responses/GeneralError'
...
  /fees:
    get:
      x-eov-operation-id: getFees
      x-eov-operation-handler: handlers
      description: Get transaction fees for coin
      parameters:
        - $ref: '#/components/parameters/deviceId'
        - $ref: '#/components/parameters/cryptoId'
      security:
        - deviceSignature: []
      responses:
        '200':
          description: Fees for coin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeList'
        default:
          $ref: '#/components/responses/GeneralError'
...
  /openalias:
    get:
      x-eov-operation-id: resolveOpenalias
      x-eov-operation-handler: handlers
      description: Resolve openalias domain
      parameters:
        - $ref: '#/components/parameters/deviceId'
        - in: query
          name: hostname
          required: true
          description: Openalias hostname
          schema:
            $ref: '#/components/schemas/Domain'
      security:
        - deviceSignature: []
      responses:
        '200':
          description: Storage
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  address:
                    $ref: '#/components/schemas/AnyCryptoAddress'
                  name:
                    $ref: '#/components/schemas/AnyString256'
                required:
                  - address
                  - name
        default:
          $ref: '#/components/responses/GeneralError'
...
  /register:
    post:
      x-eov-operation-id: register
      x-eov-operation-handler: handlers
      description: Register new device
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                walletId:
                  $ref: '#/components/schemas/Hex64String'
                deviceId:
                  $ref: '#/components/schemas/Hex64String'
                pinHash:
                  $ref: '#/components/schemas/Hex64String'
              required:
                - walletId
                - deviceId
                - pinHash
      responses:
        '201':
          description: Registration Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PublicToken'
                  - $ref: '#/components/schemas/PrivateToken'
        default:
          $ref: '#/components/responses/GeneralError'
...
  /update/{distribution}/{arch}/{version}:
    get:
      x-eov-operation-id: getUpdate
      x-eov-operation-handler: handlers
      x-allow-unknown-query-parameters: true
      description: Get release
      parameters:
        - $ref: '#/components/parameters/distribution'
        - $ref: '#/components/parameters/arch'
        - $ref: '#/components/parameters/version'
      security: []
      responses:
        '200':
          description: Release
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Update'
        default:
          $ref: '#/components/responses/GeneralError'
  /update/win/x64/{version}/RELEASES:
    get:
      x-eov-operation-id: getWinReleases
      x-eov-operation-handler: handlers
      x-allow-unknown-query-parameters: true
      description: Get release
      parameters:
        - $ref: '#/components/parameters/version'
      security: []
      responses:
        '200':
          description: Release content
          content:
            text/html:
              schema:
                type: string
                pattern: ^.+ https:\/\/.+ \d+$
                maxLength: 2048
        default:
          $ref: '#/components/responses/GeneralError'
  /updates:
    get:
      x-eov-operation-id: getUpdates
      x-eov-operation-handler: handlers
      description: Get all releases
      security: []
      responses:
        '200':
          description: Releases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateList'
        default:
          $ref: '#/components/responses/GeneralError'
  /username:
    put:
      x-eov-operation-id: setUsername
      x-eov-operation-handler: handlers
      description: Update username
      parameters:
        - $ref: '#/components/parameters/deviceId'
      security:
        - deviceSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsernameObject'
      responses:
        '200':
          description: Update username result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsernameObject'
        default:
          $ref: '#/components/responses/GeneralError'
...
components:
  schemas:
    AnyCryptoAddress:
      type: string
      pattern: ^.{1,256}$
      maxLength: 256
    AnyString256:
      type: string
      pattern: ^.{0,256}$
      maxLength: 256
    Arch:
      type: string
      pattern: ^(any|x64)$
      maxLength: 64
...
    Base64String:
      type: string
      pattern: ^(?:[A-Za-z\d+\/]{4})*?(?:[A-Za-z\d+\/]{2}(?:==)?|[A-Za-z\d+\/]{3}=?)?$
      maxLength: 2097152
      nullable: true
...
    CryptoId:
      type: string
      pattern: ^.+@.+$
      maxLength: 256
...
    CsFee:
      type: object
      additionalProperties: false
      properties:
        fee:
          type: number
          format: float
          minimum: 0
          maximum: 1
        minFee:
          $ref: '#/components/schemas/UInt64'
        maxFee:
          $ref: '#/components/schemas/UInt64'
        rbfFee:
          $ref: '#/components/schemas/UInt64'
        addresses:
          type: array
          maxItems: 128
          items:
            $ref: '#/components/schemas/AnyCryptoAddress'
        whitelist:
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/AnyCryptoAddress'
        skipMinFee:
          type: boolean
      required:
        - fee
        - minFee
        - maxFee
        - rbfFee
        - addresses
        - whitelist
        - skipMinFee
    Distribution:
      type: string
      pattern: ^[a-z0-9-]+$
      maxLength: 64
    Domain:
      type: string
      pattern: ^.+\..+$
      maxLength: 256
    Fee:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          pattern: ^(minimum|default|fastest)$
          maxLength: 64
        value:
          $ref: '#/components/schemas/UInt64'
        default:
          type: boolean
      required:
        - name
        - value
    FeeList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/Fee'
      required:
        - items
    Hex128String:
      type: string
      pattern: ^[a-f0-9]{128}$
      maxLength: 128
    Hex64String:
      type: string
      pattern: ^[a-f0-9]{64}$
      maxLength: 64
...
    PrivateToken:
      type: object
      properties:
        privateToken:
          $ref: '#/components/schemas/Hex128String'
      required:
        - privateToken
...
    PublicToken:
      type: object
      properties:
        publicToken:
          $ref: '#/components/schemas/Hex128String'
      required:
        - publicToken
    RedirectString:
      type: string
      pattern: ^Found\. Redirecting to (http|https):\/\/.+$
      maxLength: 2048
...
    UInt64:
      type: integer
      format: int64
      minimum: 0
      maximum: 9007199254740991
    Update:
      type: object
      additionalProperties: false
      properties:
        name:
          $ref: '#/components/schemas/Version'
        version:
          $ref: '#/components/schemas/Version'
        url:
          $ref: '#/components/schemas/UrlString'
        distribution:
          $ref: '#/components/schemas/Distribution'
        arch:
          $ref: '#/components/schemas/Arch'
        app:
          type: string
          pattern: ^(app|electron)$
          maxLength: 64
      required:
        - name
        - version
        - url
    UpdateList:
      type: array
      maxItems: 20
      items:
        $ref: '#/components/schemas/Update'
    UrlString:
      type: string
      pattern: ^(mailto:|(http|https):\/\/).+$
      maxLength: 2048
    Username:
      type: string
      pattern: ^[a-z0-9-]+$
      maxLength: 64
    UsernameObject:
      type: object
      additionalProperties: false
      properties:
        username:
          $ref: '#/components/schemas/Username'
      required:
        - username
    Version:
      type: string
      pattern: ^v.+$
      maxLength: 64
...
  parameters:
...
    arch:
      in: path
      name: arch
      schema:
        $ref: '#/components/schemas/Arch'
      required: true
      description: App architecture
...
    cryptoId:
      in: query
      name: crypto
      required: true
      description: Crypto ID
      allowReserved: true
      schema:
        $ref: '#/components/schemas/CryptoId'
...
    deviceId:
      in: query
      name: id
      required: true
      description: Device ID
      schema:
        $ref: '#/components/schemas/Hex64String'
    distribution:
      in: path
      name: distribution
      schema:
        $ref: '#/components/schemas/Distribution'
      required: true
      description: App distribution
...
    maybeDeviceId:
      in: query
      name: id
      required: false
      description: Optional Device ID
      schema:
        $ref: '#/components/schemas/Hex64String'
...
    version:
      in: path
      name: version
      schema:
        $ref: '#/components/schemas/Version'
      required: true
      description: App version
  responses:
...
    GeneralError:
      description: Unexpected error
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required:
              - code
              - error
            properties:
              code:
                type: integer
                format: int32
                minimum: 0
                maximum: 1000
              error:
                type: string
                pattern: ^.{0,16384}$
                maxLength: 16384
    SuccessResponse:
      description: Success Response
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              success:
                type: boolean
            required:
              - success
  securitySchemes:
    apiKey:
      type: apiKey
      in: query
      name: apikey
    deviceSignature:
      type: apiKey
      in: header
      name: Signature
    walletSignature:
      type: apiKey
      in: header
      name: Signature